Add deploy keys
This commit features the inclusion of private keys that allow for read access
to some git repositories on the svsticky organisation. Including these in this
source repository is not a big deal, since everyone that is able to see these
keys will also be able to access the git repos in question.