Move security headers to separate file
I have moved the security HTTP headers to a
separate file, to avoid duplication and make
it easier to let Rails deal with this itself
for Koala. Default config is an exception,
since in a HTTP-only config HSTS should not be
added.

Also fixed a few spelling/style errors in the
process.