Check for a dpkg lock before updating packages
The apt task that updates all packages would fail if unattended-upgrades would
happen to be running in the background, because of the dpkg lock. This happened
often when a staging droplet that was created from a snapshot, was just spun up.

This workaround should be replaced by an apt parameter when that arrives in
Ansible (#25414).