Remove duplicate SSH keys added during bootstrap
The bootstrap script creates the ansible user and copies all SSH keys from the
root user to this new user. Later on, the keys for the ansible user are re-added
and managed by the playbook itself. But it would never touch the "old" keys
added during the bootstrapping.

These tasks delete the keys at the top of the 'ansible' user's authorized_keys
that aren't preceded by the marker (comment).