Change use of task variables
I used to use a loop (with_items) to use a single variable in a task, which
isn't like it should be. Using a loop also puts the output of a registered
variable in an array, which isn't logical when the task isn't actually
"looped".

Replaced that by task-specific vars.