Only reload nginx when actually needed
nginx was reloaded every run, but I changed it to only reload when the
handler is actually notified.

The extra flushing of handlers was needed, because when the reboot handler was
executed just before Certbot was run, Certbot would fail because the kernel's
RNG was not seeded yet. So I also flushed the handlers earlier, to give the
system time to fully complete its reboot.