Remove prompt and add check in bootstrapper
The prompt to ask what server to bootstrap, was not actually used, so it would
not work if you used anything other than the default value. I removed this,
since a new hostname would have to be changed in the inventory anyway if you
wanted to deploy afterwards.

Also added a check, so it won't try to bootstrap a server that's already been
bootstrapped, by checking if the ansible user already exists.