Fix theoretical TLS certificate bug
Certificate Authorities let people prove ownership of a domain by checking the
existence of a file at
http(s)://example.com/.well-known/(acme-challenge|pki-validation)/.
This is used to get TLS certificates.

Let's Encrypt uses http://example.com/.well-known/acme-challenge/ for this
(so only port 80), so I blocked the abovementioned URL on all of our websites,
because committee users should not be able to obtain certificates for our
domains. I didn't change the existing location block that was used for Let's
Encrypt's validations, so that won't break.