Fix variable bug in Bash scripts
Fixed a bug in Bash's [parameter substitution] we used in our shell scripts.
The old construction would take the "x" value when the var *was* set, instead
of when it *wasn't* set. Since most of our tests don't use the value, this
hasn't had consequences in use. The ":" makes sure the var is also substituted
when the var is assigned a null value.

[parameter substitution]:
https://www.tldp.org/LDP/abs/html/parameter-substitution.html