Add Referer tracking for Execut websites
Adds logging of Referer headers for visits to the Execut website. Logging is
only done if:

1. The Do Not Track header is not enabled (so absent or 0), and
2. The Referer header specifies a referrer not under *.execut.nl (we only care
    about incoming visits).

All Execut websites log referrers to /var/log/nginx/execut_referer_tracking in
a Tab-seperated file containing Referer, URL, timestamp and IP. This file needs
to be either pseudonymized or anonymized before sharing the data.

And yes, Referer is a misspelling, but it is the correct header name:
https://en.wikipedia.org/wiki/HTTP_referer